Remove duplicate code by using shared component
