+++ apps/page/src/components/embed/EmbedPage.tsx
@@ -1,36 +1,5 @@
-import SlTerminal, { TrainType, type TrainTypeValue } from '../slTerminal';
+import EmbedTerminal from '../EmbedTerminal';
 
-const parseMessages = (messagesParam: string | null): string[] => {
-  try {
-    return JSON.parse(decodeURIComponent(messagesParam ?? '[]')) as string[];
-  } catch (error) {
-    console.warn('Failed to parse messages parameter:', error);
-    return [];
-  }
-};
-
-const EmbedPage = () => {
-  const searchParams = new URLSearchParams(window.location.search);
-  const accident = searchParams.get('accident') === 'true';
-  const fly = searchParams.get('fly') === 'true';
-  const smoke = searchParams.get('smoke') !== 'false';
-  const trainType = (searchParams.get('trainType') as TrainTypeValue | null) ?? TrainType.D51;
-  const messages = parseMessages(searchParams.get('messages'));
-  const fontColor = searchParams.get('fontColor') ?? '#0f0';
-  const backgroundColor = searchParams.get('backgroundColor') ?? '#000';
-
-  return (
-    <SlTerminal
-      title="SL"
-      accident={accident}
-      fly={fly}
-      smoke={smoke}
-      trainType={trainType}
-      messages={messages}
-      fontColor={fontColor}
-      backgroundColor={backgroundColor}
-    />
-  );
-};
+const EmbedPage = () => <EmbedTerminal />;
 
 export default EmbedPage;
